+++ src/main/scala/firrtl/ir/IR.scala
@@ -289,7 +289,7 @@ case class DefMemory(
     info: Info,
     name: String,
     dataType: Type,
-    depth: Int,
+    depth: BigInt,
     writeLatency: Int,
     readLatency: Int,
     readers: Seq[String],
